refactoring & renaming
