+++ src/messaging/index.ts
@@ -1,0 +1,2 @@
+export { default as MessageQueue } from './queue'
+export { default as MessageTopic } from './topic'
